requirements.txt
aiogram==2.9
aiohttp==3.7.2
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.21.0
attrs==20.3.0
Babel==2.8.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==8.1.3
colorama==0.4.4
idna==2.10
joblib==1.1.0
multidict==5.0.0
nltk==3.7
python-dotenv==0.15.0
pytz==2020.4
regex==2022.4.24
requests==2.24.0
soupsieve==2.0.1
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
